Add changes for google calendar and token storage

+++ src/app/api/googleCalendar/events/deleteEvent/route.js
@@ -3,28 +3,18 @@ import { NextResponse } from "next/server";
 import { getTokenByAppName } from "@/lib/token-connections";
 
 export async function DELETE(req) {
-  const { searchParams } = new URL(req.url);
-  const eventId = searchParams.get("eventId");
+  const { eventId } = await req.json();
   const { tokenKey: token } = await getTokenByAppName("Outlook");
 
+  if (!token || !eventId) {
+    return new NextResponse(JSON.stringify({ error: "Missing eventId" }), {
+      status: 400,
+    });
+  }
+
   try {
-    if (!token) {
-      return new NextResponse(JSON.stringify({ error: "Token is missing" }), {
-        status: 400,
-      });
-    }
-
-    if (!eventId) {
-      return new NextResponse(
-        JSON.stringify({ error: "Event ID is required" }),
-        {
-          status: 400,
-        }
-      );
-    }
-
     const response = await fetch(
-      `https://graph.microsoft.com/v1.0/me/events/${eventId}`,
+      `https://www.googleapis.com/calendar/v3/calendars/primary/events/${eventId}`,
       {
         method: "DELETE",
         headers: {
@@ -35,7 +25,6 @@ export async function DELETE(req) {
 
     if (!response.ok) {
       const errorData = await response.json();
-      console.error("Error deleting event:", errorData);
       return new NextResponse(
         JSON.stringify({ error: "Failed to delete event", details: errorData }),
         { status: response.status }
@@ -43,15 +32,12 @@ export async function DELETE(req) {
     }
 
     return new NextResponse(
-      JSON.stringify({
-        message: `Event with ID ${eventId} deleted successfully`,
-      }),
+      JSON.stringify({ message: "Event deleted successfully" }),
       { status: 200 }
     );
   } catch (error) {
-    console.error("Error deleting event:", error);
     return new NextResponse(
-      JSON.stringify({ error: "Internal server error" }),
+      JSON.stringify({ error: "Internal Server Error", details: error }),
       { status: 500 }
     );
   }
